Implement OwnedStack.
