+++ src/stack.rs
@@ -4,6 +4,7 @@
 // http://apache.org/licenses/LICENSE-2.0> or the MIT license <LICENSE-MIT or
 // http://opensource.org/licenses/MIT>, at your option. This file may not be
 // copied, modified, or distributed except according to those terms.
+//! Traits for stacks.
 
 /// A trait for objects that hold ownership of a stack.
 pub trait Stack {
@@ -22,22 +23,3 @@ pub trait Stack {
 /// A guarded stack must guarantee that any access of data at addresses `limit()` to
 /// `limit().offset(4096)` will abnormally terminate the program.
 pub unsafe trait GuardedStack {}
-
-/// SliceStack holds a non-guarded stack allocated elsewhere and provided as a mutable
-/// slice.
-pub struct SliceStack<'a>(pub &'a mut [u8]);
-
-impl<'a> Stack for SliceStack<'a> {
-    #[inline(always)]
-    fn base(&self) -> *mut u8 {
-        // The slice cannot wrap around the address space, so the conversion from usize
-        // to isize will not wrap either.
-        let len: isize = self.0.len() as isize;
-        unsafe { self.limit().offset(len) }
-    }
-
-    #[inline(always)]
-    fn limit(&self) -> *mut u8 {
-        self.0.as_ptr() as *mut u8
-    }
-}
