+++ tests/stack.rs
@@ -6,12 +6,18 @@
 // copied, modified, or distributed except according to those terms.
 extern crate fringe;
 
-use fringe::{Stack, SliceStack, OsStack};
+use fringe::{Stack, SliceStack, OwnedStack, OsStack};
 
 #[test]
 fn slice_stack() {
     let mut memory = [0; 1024];
     let stack = SliceStack(&mut memory);
+    assert_eq!(stack.base() as isize - stack.limit() as isize, 1024);
+}
+
+#[test]
+fn owned_stack() {
+    let stack = OwnedStack::new(1024);
     assert_eq!(stack.base() as isize - stack.limit() as isize, 1024);
 }
